WinManager class for managing winconditions
It makes the main code a bit cleaner and it
will be easier to add new winconditions in the future.
For example a point system or quests.

In this commit:

I also deleted the assignment operator
of the Point class, as it will be auto-generated by the compiler
either way. (I don't know why I created it in the first place.)

I also fixed a typo: INTIAL->INITIAL